Inject nREPL 0.4 instead of nREPL 0.2
The only downside of that change is that users of older Boot/Lein
will end up using an older version of the legacy nREPL 0.2, but that
should not affect them negatively, as the differences between 0.2.12
and 0.2.13 are pretty small.

Generally we could have just injected both versions, but I've noticed
that's pretty confusing for the users, so I'd rather us not do it.