[Fix #2466] Make generic cider ops use any available nrepl connection
Make "any" a synonym of "multi" for `cider-current-repl`, which makes
it more clear any avaiable nrepl connection can be used.

This allows ops for like documentatation browser to work, even if only a repl
for a different type of repl is connected. For example, when formatting a .cljs
buffer, a connected "clj" repl will suffice.

Cider eval uses a separate repl connection selection, which lets it unaffacted
by these changes.