More precise point positioning for cider-format-defun and no mark
cider-format-region uses the current point position to figure out where to
place the point after formatting. mark-defun manipulates point, so if it was in
the middle of some defun, after formatting the point would be placed at the
beginning of the defun.

Additonally, by not marking any region, it is not necessary to clean up, which
previously was broken since save-excursion didn't restore mark since Emacs 25.1