Fix restoring point after formatting in some conditions
The snippet is used as a regex, so it needs to be quoted to not cause any regex
errors further down and also make the search more precise.

An example would be that the snippet would contain the opening [ of a clojure
vector, but not the closing ], which lead to an invalid regex and thus make
re-search-forward/re-search-backward break.